libsel4 with no libc dependency.
The primary changes are introducing sel4_types.h and removing std* types
plus porting assert and printf code from the kernel to libsel4. All of
this means the code within libsel4 does not overload any typical libc
entities. So now libsel4 uses types like seL4_Uint32 ... instead of
uint32_t. And printf is now seL4_Printf and assert is seL4_Assert ....

I'm also using sel4_ prefixes for various files as I felt it was more
consistent with the names of the entities within the files.

The only new library is libsel4_benchmark and since it consists of just
sel4_benchmark.h we might want to move that back into libsel4. I would
have liked to move out libsel4_assert, libsel4_printf and libsel4_putchar
but since asserts are used by low level generated code I couldn't come up
with a good way of doing that.

Finally, the only file modified that effects kernel code is
kernel/tools/bitfield_gen.py. It needed to be modified as it generates
files for both kernel and user space. And for user space the generated code
(types_gen.h) needed to use the new types and asserts. The changes should
not change what is generated for the kernel and I did a comparison of
kernel_final.{c|s} before and after my change and the only differences
were time stamps